added GA tests to Prod
